Sped up telegram group and channel creation. Also integrated schedule messages

+++ CLE/Module_CommunicationManagement/views.py
@@ -1,7 +1,9 @@
+import json
 import time
 import traceback
 from datetime import datetime, timedelta
 from django.shortcuts import render
+from django.http import HttpResponse
 from Module_TeamManagement.models import *
 from telethon.tl.types import Channel, Chat
 from Module_Account.src import processLogin
@@ -66,7 +68,8 @@ def faculty_telegram_Base(requests,response=None):
             if tele_util.dialogExists(client,telegram_chat.name,tele_chat_type[telegram_chat.type]):
                 response['telegram_chats'].append({'name': telegram_chat.name})
             else:
-                telegram_chat.delete()
+                if telegram_chat.link != None:
+                    telegram_chat.delete()
 
         if len(response['telegram_chats']) > 0:
             if telegram_chat_name == None:
@@ -126,6 +129,7 @@ def faculty_telegram_UpdateChatMembers(requests):
     response['message'] = 'Members successfully updated'
     return faculty_telegram_Base(requests,response)
 
+
 # Group creation form
 #
 def faculty_telegram_CreateGroup(requests):
@@ -146,26 +150,23 @@ def faculty_telegram_CreateGroup(requests):
     print('Additional Username: ' + additional_username)
 
     try:
-        username = requests.user.email.split('@')[0]
-        client = tele_util.getClient(username)
-
-        results = tele_util.initialize_Group(
-            username=additional_username,
-            client=client,
-            group_name=group_name,
-        )
-
         # Create Telegram_Chats object
         try:
             telegram_chat = Telegram_Chats.objects.create(
                 name=group_name,
                 type='Group',
-                link=results['group_link'],
-                members=None,
             )
             telegram_chat.save()
         except:
             telegram_chat = Telegram_Chats.objects.get(name=group_name)
+
+        username = requests.user.email.split('@')[0]
+        tasks.createChannel(
+            username=username,
+            group_name=group_name,
+            additional_username=additional_username,
+            schedule=0,
+        )
 
         # Assign to the students of the course_section
         class_QuerySet = Class.objects.filter(course_section=course_section)
@@ -203,33 +204,28 @@ def faculty_telegram_CreateChannel(requests):
     print('Course Section: ' + course_section)
 
     try:
-        username = requests.user.email.split('@')[0]
-        client = tele_util.getClient(username)
-
-        results = tele_util.initialize_Channel(
-            client=client,
-            channel_name=channel_name,
-        )
-
         # Create Telegram_Chats object
         try:
             telegram_chat = Telegram_Chats.objects.create(
                 name=channel_name,
                 type='Channel',
-                link=results['channel_link'],
-                members=None,
             )
             telegram_chat.save()
         except:
             telegram_chat = Telegram_Chats.objects.get(name=channel_name)
+
+        username = requests.user.email.split('@')[0]
+        tasks.createChannel(
+            username=username,
+            channel_name=channel_name,
+            schedule=0,
+        )
 
         # Assign to the students of the course_section
         class_QuerySet = Class.objects.filter(course_section=course_section)
         for student in class_QuerySet:
             student.telegram_chats.add(telegram_chat)
             student.save()
-
-        tele_util.disconnectClient(client)
 
     except Exception as e:
         traceback.print_exc()
@@ -288,3 +284,24 @@ def faculty_telegram_SendMessage(requests):
 
     response['message'] = 'Message successfully sent'
     return faculty_telegram_Base(requests,response)
+
+
+def faculty_telegram_GetChatLink(requests):
+    response = { 'faculty_telegram_GetChatLink' : 'active'}
+
+    # Redirect user to login page if not authorized and faculty
+    try:
+        processLogin.InstructorVerification(requests)
+    except:
+        logout(requests)
+        return render(requests,'Module_Account/login.html',response)
+
+    telegram_chat_name = requests.POST.get('telegram_chat_name')
+
+    try:
+        telegram_chatObj = Telegram_Chats.objects.get(name=telegram_chat_name)
+        response['telegram_chat_link'] = telegram_chatObj.link
+    except:
+        traceback.print_exc()
+
+    return HttpResponse(json.dumps(response), content_type='application/json')
